+++ tests/Pest.php
@@ -13,4 +13,4 @@
 
 use Tailpipe\Tests\TestCase;
 
-uses(TestCase::class)->in('Feature');
+uses(TestCase::class)->in('Unit');