+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.1.0
+## 1.3.0
+- Added the `opacity` variable to the default list of variables to process
+- Added the `parse: true` option that allows outputting ["parsed values"](README.md#parsed-values)
+- Added a real-world example to the README
+
+## 1.2.0
 
 - Added the `fontFamily` variable to the default list of variables to process